Finally get door state asserting working
This commit adds functionality to open or close a door. The door
managed will be the first one listed in your devices. You can open
and close the door, as well as report on the door state. Examples
of how this can work are provided in the `door.php` script.